feature: Merge new telemetry into Printer state.

+++ PrinterService.Host/PrusaConnect/DTO/Telemetry/EnclosureTelemetryDTO.cs
@@ -2,11 +2,16 @@ using System.Text.Json.Serialization;
 
 namespace PrinterService.Host.PrusaConnect.DTO.Telemetry;
 
+/// <summary>
+/// Wire values are all integers - unlike <see cref="ChamberTelemetryDTO.Temperature"/>, which is
+/// fixed-point. Confirmed against firmware (<c>render.cpp</c>): the enclosure block renders with
+/// <c>JSON_FIELD_INT</c> throughout, versus chamber's <c>JSON_FIELD_FFIXED</c> for temperature.
+/// </summary>
 public class EnclosureTelemetryDTO
 {
     [JsonPropertyName("temp")]
-    public float Temperature { get; set; }
-    
+    public int Temperature { get; set; }
+
     [JsonPropertyName("fan_rpm")]
     public int FanSpeed { get; set; }
